+++ test/PeriodTest.php
@@ -10,7 +10,7 @@ use DateTimeZone;
 use Generator;
 use JsonSerializable;
 use League\Period\Period;
-use PHPUnit_Framework_TestCase as TestCase;
+use PHPUnit\Framework\TestCase;
 
 class ExtendedDate extends DateTimeImmutable
 {